bugfix: Fix circular dependency. Provide single source for AppProviders.
